Remove mongomock (#5)
Successfully remove mongomock for local testing.

Start testing POST endpoint.
Ensure OAuth2 authentication can be disabled when testing. This is
important, since we want to test against a live backend, i.e., a
configuration option should be toggleable as to enable/disable OAuth2
auth.

Update line-length to 108.

Support testing users with different priviliges and testing supplying
OAuth credentials or not.

Update tests to support as many tests as possible with a live backend.