Fix test user creation
The basic auth users must be created in the 'admin' database (by default
it was using the 'test' database).

Furthermore, mongo-express is added as a service for easier access into
the content of the test MongoDB.

A `test_token` configuration is also added to be used if
`external_oauth` is `False`.

Finally, Python 3.12 is used in the Dockerfile.