Update codebase to use the soft7 package
As a "breaking" change, this means the `uri` (or not `identity`) value
is no longer so strictly guarded and validated. There is now no error if
the namespace or composition is malformed according to the TEAM4.0-style
URLs.

The ValidationError outputs should also be updated in the
tests/static/invalid_entities/README.md file.