Remove the `service` module
Everything under this module is moved to the root package directory.

Instead of initializing the server settings directly in the config
module, it is now wrapped in an LRU cached function.
The code base has been updated to use this new way of retrieving the
server configuration.

Instead of creating the FastAPI application directly in the main module,
it is instead wrapped in a function.
THe code base has been updated to use this new way of retrieving the
application.
Especially, a new `asgi.py` file has been added to the root of the
repository, to be used with the Dockerfile.