Remove X509 user and self-signed certificates
Remove all things to do with the extra security around the MongoDB
server and the split in a write and read-only user. Simply use the same
write-access user for everything on a basic username/password
authentication level.

This is because the server is not intended to be exposed outside the
internal network.

Also, remove the logging module as this is handled by the main
DataSpaces server.

Furthermore, move to use uvicorn-worker instead of uvicorn.